+++ packages/order-signing/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## [1.1.13](https://github.com/cowprotocol/cow-sdk/compare/sdk-order-signing-v1.1.12...sdk-order-signing-v1.1.13) (2026-09-25)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @cowprotocol/sdk-common bumped to 0.14.1
+    * @cowprotocol/sdk-config bumped to 2.7.1
+    * @cowprotocol/sdk-contracts-ts bumped to 3.7.1
+    * @cowprotocol/sdk-order-book bumped to 4.1.1
+  * devDependencies
+    * @cowprotocol/sdk-ethers-v5-adapter bumped to 0.4.17
+    * @cowprotocol/sdk-ethers-v6-adapter bumped to 0.4.17
+    * @cowprotocol/sdk-viem-adapter bumped to 0.3.32
+
 ## [1.1.12](https://github.com/cowprotocol/cow-sdk/compare/sdk-order-signing-v1.1.11...sdk-order-signing-v1.1.12) (2026-09-21)